Add highlighter to editor in config-server
